클릭 이벤트 처리

+++ src/main/java/com/rumor/trash/repository/TrashRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrashRepository extends JpaRepository<Trash, Integer> {
+public interface TrashRepository extends JpaRepository<Trash, Long> {
 }
